refactor(#4236): move shared enums and guards into core domains
