+++ ui/cypress/support/builder/UserBuilder.ts
@@ -17,7 +17,7 @@
  */
 
 import { User } from '../model/User';
-import { UserRole } from '../../../src/app/_enums/user-role.enum';
+import { UserRole } from '../../../src/app/core/auth/user-role.enum';
 
 export class UserBuilder {
     user: User;
