+++ ui/src/app/chart/components/chart-overview/chart-overview.component.ts
@@ -26,7 +26,7 @@ import { AuthService } from '../../../services/auth.service';
 import { SpChartRoutes } from '../../chart.breadcrumb';
 import { ChartRoutingService } from '../../../chart-shared/services/chart-routing.service';
 import { ChartOverviewTableComponent } from './chart-overview-table/chart-overview-table.component';
-import { UserPrivilege } from '../../../_enums/user-privilege.enum';
+import { UserPrivilege } from '../../../core/auth/user-privilege.enum';
 import { Subscription } from 'rxjs';
 import {
     FlexDirective,
