+++ ui/src/app/configuration/sites-configuration/sites-configuration.component.ts
@@ -29,7 +29,7 @@ import {
     SpNavigationItem,
 } from '@streampipes/shared-ui';
 import { SpConfigurationRoutes } from '../configuration.breadcrumb';
-import { UserRole } from '../../_enums/user-role.enum';
+import { UserRole } from '../../core/auth/user-role.enum';
 import { LayoutDirective } from '@ngbracket/ngx-layout/flex';
 import { LocationFeaturesConfigurationComponent } from './location-features-configuration/location-features-configuration.component';
 import { SiteAreaConfigurationComponent } from './site-area-configuration/site-area-configuration.component';
