+++ ui/src/app/core/auth/guards/auth.can-activate-children.guard.ts
@@ -16,31 +16,24 @@
  *
  */
 
+import { Injectable } from '@angular/core';
 import {
     ActivatedRouteSnapshot,
-    CanActivate,
     CanActivateChild,
     GuardResult,
     MaybeAsync,
     RouterStateSnapshot,
 } from '@angular/router';
-import { AuthService } from '../services/auth.service';
-import { Injectable } from '@angular/core';
+import { AuthService } from '../../../services/auth.service';
 
 @Injectable({ providedIn: 'root' })
-export class PageAuthGuard implements CanActivate, CanActivateChild {
+export class AuthCanActivateChildrenGuard implements CanActivateChild {
     constructor(private authService: AuthService) {}
 
-    canActivate(
-        route: ActivatedRouteSnapshot,
+    canActivateChild(
+        childRoute: ActivatedRouteSnapshot,
         state: RouterStateSnapshot,
     ): MaybeAsync<GuardResult> {
-        return this.canActivateChild(route);
-    }
-
-    canActivateChild(activatedRouteSnapshot: ActivatedRouteSnapshot): boolean {
-        const privileges: string[] = activatedRouteSnapshot.data.privileges;
-
-        return this.authService.isAnyAccessGranted(privileges, true);
+        return this.authService.ensureAuthenticated(state.url);
     }
 }
