+++ ui/src/app/core/navigation/page-name.enum.ts
