+++ ui/src/app/core/setup/guards/base-configured.can-activate.guard.ts
@@ -24,7 +24,7 @@ import {
     UrlTree,
 } from '@angular/router';
 import { Observable } from 'rxjs';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../../../services/auth.service';
 
 export abstract class BaseConfiguredCanActivateGuard implements CanActivate {
     constructor(
