+++ ui/src/app/core/setup/guards/configured.can-activate.guard.ts
@@ -18,7 +18,7 @@
 
 import { Injectable } from '@angular/core';
 import { Router, UrlTree } from '@angular/router';
-import { AuthService } from '../services/auth.service';
+import { AuthService } from '../../../services/auth.service';
 import { BaseConfiguredCanActivateGuard } from './base-configured.can-activate.guard';
 
 @Injectable({ providedIn: 'root' })
