+++ ui/src/app/home/components/status.component.ts
@@ -20,7 +20,7 @@ import { Component, HostBinding, inject, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { AssetLinkType, UserInfo } from '@streampipes/platform-services';
 import { StatusBox } from '../models/home.model';
-import { UserRole } from '../../_enums/user-role.enum';
+import { UserRole } from '../../core/auth/user-role.enum';
 import { zip } from 'rxjs';
 import { TranslatePipe } from '@ngx-translate/core';
